Added subfolders to XmlProcessors (rssimport and xmlexport) to reflect the future project-structure
