+++ XmlProcessor/XmlProcessor.csproj
@@ -41,17 +41,20 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DeserializingManager.cs" />
-    <Compile Include="XmlLoader.cs" />
-    <Compile Include="EpisodeMappedToXml.cs" />
+    <Compile Include="RssImport\DeserializingManager.cs" />
+    <Compile Include="RssImport\XmlLoader.cs" />
+    <Compile Include="RssImport\EpisodeMappedToXml.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SeriesMappedToXml.cs" />
+    <Compile Include="RssImport\SeriesMappedToXml.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\CommonTypes\CommonTypes.csproj">
       <Project>{014FB56A-6592-44D1-ABBC-36C4178694FE}</Project>
       <Name>CommonTypes</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="XmlExport\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
